requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.3.1+cpu
torchaudio==2.3.1+cpu
transformers==4.57.1
faster-whisper==1.2.0
pydub
f5-tts==1.1.17
moviepy==2.0.0
firebase_admin
imagekitio==4.2.0
python-dotenv
pandas